Changing barcode to long across program

+++ src/main/java/CheckItemsController/ControllerCheckoutPage.java
@@ -315,7 +315,7 @@ public class ControllerCheckoutPage extends ControllerMenu implements IControlle
      */
     private void submitMultipleItems() {
         Student thisStudent = database.selectStudent(getstudentID());
-        List<Integer> barcodes = new ArrayList<>();
+        List<Long> barcodes = new ArrayList<>();
         if (barcodeIsNotEmpty(barcode)) {
             barcodes.add(getBarcode());
         }
@@ -374,7 +374,7 @@ public class ControllerCheckoutPage extends ControllerMenu implements IControlle
      *
      * @return True if item is being checked in
      */
-    private boolean itemIsBeingCheckedIn(int barcode) {
+    private boolean itemIsBeingCheckedIn(long barcode) {
         checkoutParts = checkOut.returnCheckedOutObjects();
         int studentID = getstudentID();
 
@@ -395,7 +395,7 @@ public class ControllerCheckoutPage extends ControllerMenu implements IControlle
      *
      * @return True if item is extended
      */
-    private boolean extendedCheckoutIsSelected(int barcode) {
+    private boolean extendedCheckoutIsSelected(long barcode) {
         return !itemIsBeingCheckedIn(barcode) && extended.isSelected();
     }
 
@@ -404,7 +404,7 @@ public class ControllerCheckoutPage extends ControllerMenu implements IControlle
      *
      * @return True if item is faulty
      */
-    private boolean itemBeingCheckedBackInIsFaulty(int barcode) {
+    private boolean itemBeingCheckedBackInIsFaulty(long barcode) {
         return itemIsBeingCheckedIn(barcode) && faulty.isSelected();
     }
 
@@ -634,9 +634,9 @@ public class ControllerCheckoutPage extends ControllerMenu implements IControlle
      *
      * @return barcode as integer
      */
-    private int getBarcode() {
+    private long getBarcode() {
         if(!barcode.getText().isEmpty()) {
-            return Integer.parseInt(barcode.getText());
+            return Long.parseLong(barcode.getText());
         }
         return 0;
     }
@@ -645,9 +645,9 @@ public class ControllerCheckoutPage extends ControllerMenu implements IControlle
      * Gets barcode as text, returns as int
      * @return barcode as integer
      */
-    private int getBarcode2() {
+    private long getBarcode2() {
         if(!barcode2.getText().isEmpty()) {
-            return Integer.parseInt(barcode2.getText());
+            return Long.parseLong(barcode2.getText());
         }
         return 0;
     }
@@ -657,9 +657,9 @@ public class ControllerCheckoutPage extends ControllerMenu implements IControlle
      *
      * @return barcode as integer
      */
-    private int getBarcode3() {
+    private long getBarcode3() {
         if(!barcode3.getText().isEmpty()) {
-            return Integer.parseInt(barcode3.getText());
+            return Long.parseLong(barcode3.getText());
         }
         return 0;
     }
@@ -669,9 +669,9 @@ public class ControllerCheckoutPage extends ControllerMenu implements IControlle
      *
      * @return barcode as integer
      */
-    private int getBarcode4() {
+    private long getBarcode4() {
         if(!barcode4.getText().isEmpty()) {
-            return Integer.parseInt(barcode4.getText());
+            return Long.parseLong(barcode4.getText());
         }
         return 0;
     }
@@ -681,9 +681,9 @@ public class ControllerCheckoutPage extends ControllerMenu implements IControlle
      *
      * @return barcode as integer
      */
-    private int getBarcode5() {
+    private long getBarcode5() {
         if(!barcode5.getText().isEmpty()) {
-            return Integer.parseInt(barcode5.getText());
+            return Long.parseLong(barcode5.getText());
         }
         return 0;
     }
